Return 3D segments in `Approx`
